pyproject.toml
[project]
name = "onebot-forward"
version = "0.1.0"
description = "OneBot WebSocket Forward Service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.10.5",
    "pyyaml>=6.0.2",
    "websockets>=14.1",
]
